Add acl maintenance APIs
