+++ packages/gateway-lib/src/utils/catalog.ts
@@ -5,23 +5,29 @@ import { HttpLink } from 'apollo-link-http';
 import { Request, Response } from 'express';
 import {
   DocumentNode,
-  DirectiveDefinitionNode,
-  EnumTypeDefinitionNode,
-  InputObjectTypeDefinitionNode,
-  InterfaceTypeDefinitionNode,
-  ObjectTypeDefinitionNode,
-  ScalarTypeDefinitionNode,
-  SchemaDefinitionNode,
-  UnionTypeDefinitionNode,
-  FieldDefinitionNode,
-  InputValueDefinitionNode,
-  OperationTypeDefinitionNode,
-  TypeNode,
+  // DirectiveDefinitionNode,
+  // EnumTypeDefinitionNode,
+  // InputObjectTypeDefinitionNode,
+  // InterfaceTypeDefinitionNode,
+  // ObjectTypeDefinitionNode,
+  // ScalarTypeDefinitionNode,
+  // SchemaDefinitionNode,
+  // UnionTypeDefinitionNode,
+  // FieldDefinitionNode,
+  // InputValueDefinitionNode,
+  // OperationTypeDefinitionNode,
+  // TypeNode,
 } from 'graphql';
 import gql from 'graphql-tag';
 import nodeFetch from 'node-fetch';
 import { ServiceType } from '../types';
 import { getLogger } from './getLogger';
+import {
+  checkDesc,
+  findDataType,
+  buildObjectType,
+  ANNO_IGNORE,
+} from './schemaUtils';
 
 const fetch = nodeFetch as any;
 const logger = getLogger('[gw-lib] catalog.js');
@@ -29,9 +35,9 @@ const logger = getLogger('[gw-lib] catalog.js');
 const ROOT_OPS_QUERY = 'query';
 const ROOT_OPS_MUTTN = 'mutation';
 const ROOT_OPS_SBSCP = 'subscription';
-const ANNO_SCHEMA = '@SCHEMA ';
-const ANNO_PRIMRY = '@PRIMARY';
-const ANNO_IGNORE = '@SKIP';
+// const ANNO_SCHEMA = '@SCHEMA ';
+// const ANNO_PRIMRY = '@PRIMARY';
+// const ANNO_IGNORE = '@SKIP';
 
 export const buildCatalogedSchema = (service: string, serviceType: ServiceType, enabled: boolean, sdl: {
   typeDefs: DocumentNode;
@@ -44,150 +50,156 @@ export const buildCatalogedSchema = (service: string, serviceType: ServiceType, 
 
   const srvType = (serviceType === ServiceType.Public) ? 'Public' : (serviceType === ServiceType.Private) ? 'Private' : 'Remote';
 
-  const parseType = (t: TypeNode) => {
-    let type = t;
-    let isNull = true;
-    let isList = false;
-    let cnt = 10;
-    while (type.kind !== 'NamedType' && cnt > 0) {
-      if (type.kind === 'NonNullType') {
-        isNull = false;
-      } else if (type.kind === 'ListType') {
-        isList = true;
-      }
-      type = type.type;
-      cnt --;
-    }
+  // const parseType = (t: TypeNode) => {
+  //   let type = t;
+  //   let isNull = true;
+  //   let isList = false;
+  //   let cnt = 10;
+  //   while (type.kind !== 'NamedType' && cnt > 0) {
+  //     if (type.kind === 'NonNullType') {
+  //       isNull = false;
+  //     } else if (type.kind === 'ListType') {
+  //       isList = true;
+  //     }
+  //     type = type.type;
+  //     cnt --;
+  //   }
 
-    let isPrimitive = true;
-    if (type.kind === 'NamedType') {
-      switch (type.name.value) {
-        case 'Int':
-        case 'Float':
-        case 'String':
-        case 'Boolean':
-        case 'ID':
-          isPrimitive = true;
-          break;
-        default:
-          isPrimitive = false;
-          break;
-      }
-      return { dataType: type.name.value, isPrimitive, isList, isNull };
-    } else {
-      return {};
-    }
-  };
+  //   let isPrimitive = true;
+  //   if (type.kind === 'NamedType') {
+  //     switch (type.name.value) {
+  //       case 'Int':
+  //       case 'Float':
+  //       case 'String':
+  //       case 'Boolean':
+  //       case 'ID':
+  //         isPrimitive = true;
+  //         break;
+  //       default:
+  //         isPrimitive = false;
+  //         break;
+  //     }
+  //     return { dataType: type.name.value, isPrimitive, isList, isNull };
+  //   } else {
+  //     return {};
+  //   }
+  // };
 
-  // Find data type of field
-  const findDataType = (f: FieldDefinitionNode | InputValueDefinitionNode | OperationTypeDefinitionNode) => {
-    const { dataType, isPrimitive, isList, isNull } = parseType(f.type);
-    if (dataType) {
-      if (f.kind === 'OperationTypeDefinition') {
-        return { field: { operation: f.operation }, dataType, isPrimitive };
-      } else {
-        const field = { [f.name.value]: { type: dataType }};
-        if ((checkDesc(f) & 1) > 0) field[f.name.value]['description'] = f.description.value;
-        if (!isNull)      field[f.name.value]['required'] = true;
-        if (isList)       field[f.name.value].type = `${dataType}[]`;
-        if (!isPrimitive) field[f.name.value]['ref'] = dataType.toLowerCase();
-        return { field, dataType, isPrimitive };
-      }
-    }
-    return {};
-  };
+  // // Find data type of field
+  // const findDataType = (f: FieldDefinitionNode | InputValueDefinitionNode | OperationTypeDefinitionNode) => {
+  //   const { dataType, isPrimitive, isList, isNull } = parseType(f.type);
+  //   if (dataType) {
+  //     if (f.kind === 'OperationTypeDefinition') {
+  //       return { field: { operation: f.operation }, dataType, isPrimitive };
+  //     } else {
+  //       const field = { [f.name.value]: { type: dataType }};
+  //       if ((checkDesc(f) & 1) > 0) field[f.name.value]['description'] = f.description.value;
+  //       if (!isNull)      field[f.name.value]['required'] = true;
+  //       if (isList)       field[f.name.value].type = `${dataType}[]`;
+  //       if (!isPrimitive) field[f.name.value]['ref'] = dataType.toLowerCase();
+  //       return { field, dataType, isPrimitive };
+  //     }
+  //   }
+  //   return {};
+  // };
 
-  // Check if given object contain description
-  // return 000
-  //        ││└─ 0: no description;     1: description found
-  //        │└── 0: normal types;       1: main types
-  //        └─── 0: normal description; 1: schema description
-  const checkDesc = (n: any) => {
-    if (!n['description'] || !n['description']['kind'] || n['description']['kind'] !== 'StringValue' || !n['description']['value']) {
-      return 0;
-    } else if (n['kind'] && n['kind'] === 'SchemaDefinition') {
-      if (!schemaDesc) {
-        schemaDesc = n['description']['value'];
-      }
-      return 4;
-    } else if (n['description']['value'].toUpperCase().startsWith(ANNO_SCHEMA)) {
-      if (!schemaDesc) {
-        schemaDesc = n['description']['value'].substring(8);
-      }
-      return 4;
-    } else if (n['description']['value'].toUpperCase().startsWith(ANNO_PRIMRY)) {
-      return 3;
-    } else {
-      return 1;
-    }
-  };
+  // // Check if given object contain description
+  // // return 000
+  // //        ││└─ 0: no description;     1: description found
+  // //        │└── 0: normal types;       1: main types
+  // //        └─── 0: normal description; 1: schema description
+  // const checkDesc = (n: any) => {
+  //   if (!n['description'] || !n['description']['kind'] || n['description']['kind'] !== 'StringValue' || !n['description']['value']) {
+  //     return 0;
+  //   } else if (n['kind'] && n['kind'] === 'SchemaDefinition') {
+  //     if (!schemaDesc) {
+  //       schemaDesc = n['description']['value'];
+  //     }
+  //     return 4;
+  //   } else if (n['description']['value'].toUpperCase().startsWith(ANNO_SCHEMA)) {
+  //     if (!schemaDesc) {
+  //       schemaDesc = n['description']['value'].substring(8);
+  //     }
+  //     return 4;
+  //   } else if (n['description']['value'].toUpperCase().startsWith(ANNO_PRIMRY)) {
+  //     return 3;
+  //   } else {
+  //     return 1;
+  //   }
+  // };
 
-  const buildObjectType = (d: (
-    DirectiveDefinitionNode | EnumTypeDefinitionNode | InputObjectTypeDefinitionNode | InterfaceTypeDefinitionNode |
-    ObjectTypeDefinitionNode | ScalarTypeDefinitionNode | SchemaDefinitionNode| UnionTypeDefinitionNode
-  ), i: boolean) => {
-    let included = i; // if 'i' is true, will include regardless of having comments or not
-    const found: string[] = [];
-    const chkDesc = checkDesc(d);
-    const hasDesc = (chkDesc & 1) > 0;
-    const isMain = (chkDesc & 2) > 0;
-    if (hasDesc) included = true; // Include if the type has comment
+  // const buildObjectType = (d: (
+  //   DirectiveDefinitionNode | EnumTypeDefinitionNode | InputObjectTypeDefinitionNode | InterfaceTypeDefinitionNode |
+  //   ObjectTypeDefinitionNode | ScalarTypeDefinitionNode | SchemaDefinitionNode| UnionTypeDefinitionNode
+  // ), i: boolean) => {
+  //   let included = i; // if 'i' is true, will include regardless of having comments or not
+  //   const found: string[] = [];
+  //   const chkDesc = checkDesc(d);
+  //   const hasDesc = (chkDesc & 1) > 0;
+  //   const isMain = (chkDesc & 2) > 0;
+  //   if (hasDesc) included = true; // Include if the type has comment
 
-    const fields = {};
-    const types = {};
-    if (d.kind !== 'ScalarTypeDefinition' && d.kind !== 'UnionTypeDefinition' && d.kind !== 'SchemaDefinition' && 
-        d.kind !== 'EnumTypeDefinition' && d.kind !== 'DirectiveDefinition') { // All definitions with 'fields'
-      for (const f of d.fields) {
-        if (f.kind === 'FieldDefinition') {
-          if ((checkDesc(f) & 1) > 0) included = true; // Also include if a field of the type has comment
+  //   const fields = {};
+  //   const types = {};
+  //   if (d.kind !== 'ScalarTypeDefinition' && d.kind !== 'UnionTypeDefinition' && d.kind !== 'SchemaDefinition' && 
+  //       d.kind !== 'EnumTypeDefinition' && d.kind !== 'DirectiveDefinition') { // All definitions with 'fields'
+  //     for (const f of d.fields) {
+  //       if (f.kind === 'FieldDefinition') {
+  //         if ((checkDesc(f) & 1) > 0) included = true; // Also include if a field of the type has comment
 
-          // Find base type of field
-          const { field, dataType, isPrimitive } = findDataType(f);
-          if (dataType) {
-            if (!isPrimitive) found.push(dataType); // Remember this type for the subsequent passes
-            Object.assign(fields, field);
-          }
-        }
-      }
-    } else if (d.kind === 'UnionTypeDefinition') {
-      for (const t of d.types) {
-        const { dataType, isPrimitive, isNull, isList } = parseType(t);
-        const type = { [dataType]: {}};
-        if (!isPrimitive) {
-          found.push(dataType);
-          type[dataType]['ref'] = dataType.toLowerCase();
-        }
-        if (!isNull) type[dataType]['required'] = true;
-        if (isList)  type[dataType]['isList'] = true;
-        Object.assign(types, type);
-      }
-    }
+  //         // Find base type of field
+  //         const { field, dataType, isPrimitive } = findDataType(f);
+  //         if (dataType) {
+  //           if (!isPrimitive) found.push(dataType); // Remember this type for the subsequent passes
+  //           Object.assign(fields, field);
+  //         }
+  //       }
+  //     }
+  //   } else if (d.kind === 'UnionTypeDefinition') {
+  //     for (const t of d.types) {
+  //       const { dataType, isPrimitive, isNull, isList } = parseType(t);
+  //       const type = { [dataType]: {}};
+  //       if (!isPrimitive) {
+  //         found.push(dataType);
+  //         type[dataType]['ref'] = dataType.toLowerCase();
+  //       }
+  //       if (!isNull) type[dataType]['required'] = true;
+  //       if (isList)  type[dataType]['isList'] = true;
+  //       Object.assign(types, type);
+  //     }
+  //   }
 
-    let result;
-    if (included) {
-      if (Object.keys(fields).length > 0) {
-        result = hasDesc ? {
-          description: (isMain) ? d.description.value.substring(8) : d.description.value, // type with comment given
-          fields
-        } : {
-          fields // type with no comment given
-        };
-      } else if (Object.keys(types).length > 0) {
-        result = hasDesc ? {
-          description: (isMain) ? d.description.value.substring(8) : d.description.value, // union type with comment given
-          types
-        } : {
-          types // union type with no comment given
-        };
-      } else if (hasDesc) {
-        result = { description: (isMain) ? d.description.value.substring(8) : d.description.value };
-      } else {
-        result = {}; // No comment, and no field, but somehow asked to include in the catalog...
-      }
-      if (isMain) result['main'] = true;
-      return { result, found };
-    } else {
-      return { result: undefined, found: undefined };
+  //   let result;
+  //   if (included) {
+  //     if (Object.keys(fields).length > 0) {
+  //       result = hasDesc ? {
+  //         description: (isMain) ? d.description.value.substring(8) : d.description.value, // type with comment given
+  //         fields
+  //       } : {
+  //         fields // type with no comment given
+  //       };
+  //     } else if (Object.keys(types).length > 0) {
+  //       result = hasDesc ? {
+  //         description: (isMain) ? d.description.value.substring(8) : d.description.value, // union type with comment given
+  //         types
+  //       } : {
+  //         types // union type with no comment given
+  //       };
+  //     } else if (hasDesc) {
+  //       result = { description: (isMain) ? d.description.value.substring(8) : d.description.value };
+  //     } else {
+  //       result = {}; // No comment, and no field, but somehow asked to include in the catalog...
+  //     }
+  //     if (isMain) result['main'] = true;
+  //     return { result, found };
+  //   } else {
+  //     return { result: undefined, found: undefined };
+  //   }
+  // };
+
+  const sideEffect = (v: any) => {
+    if (!schemaDesc) {
+      schemaDesc = v;
     }
   };
 
@@ -210,7 +222,7 @@ export const buildCatalogedSchema = (service: string, serviceType: ServiceType, 
         if (d.kind === 'SchemaDefinition') {
           for (const o of d.operationTypes) {
             if (o.kind === 'OperationTypeDefinition') {
-              const { dataType } = findDataType(o);
+              const { dataType } = findDataType(o, sideEffect);
               switch (o.operation) {
                 case 'query':
                   roQuery = dataType;
@@ -233,7 +245,7 @@ export const buildCatalogedSchema = (service: string, serviceType: ServiceType, 
 
         // if (types[typeName] > 0) continue; // type already processed
 
-        const { result, found } = buildObjectType(d, false);
+        const { result, found } = buildObjectType(d, false, sideEffect);
         if (result) {
           count ++;
           catalog[typeName] = result;
@@ -255,7 +267,7 @@ export const buildCatalogedSchema = (service: string, serviceType: ServiceType, 
           d.kind === 'ObjectTypeDefinition' || d.kind === 'ScalarTypeDefinition' || d.kind === 'UnionTypeDefinition' ||
           d.kind === 'DirectiveDefinition'
         ) ? d : undefined).filter(d => !!d && types[d.name.value] === 0)) {
-          const { result, found } = buildObjectType(d, true);
+          const { result, found } = buildObjectType(d, true, sideEffect);
           if (result) {
             count ++;
             pass ++;
@@ -278,10 +290,10 @@ export const buildCatalogedSchema = (service: string, serviceType: ServiceType, 
         // Always use the default name of the root operations to simplify logic when building the resuling markdown doc
         const ops = (d.name.value === roQuery) ? ROOT_OPS_QUERY : (d.name.value === roMutation) ? ROOT_OPS_MUTTN : ROOT_OPS_SBSCP;
 
-        checkDesc(d);
+        checkDesc(d, sideEffect);
         for (const f of d.fields) {
           if (f.kind === 'FieldDefinition') {
-            const { field, dataType, isPrimitive } = findDataType(f);
+            const { field, dataType, isPrimitive } = findDataType(f, sideEffect);
             if (!catalog[ops]) catalog[ops] = {};
 
             if (field[f.name.value]['description']) {
@@ -295,7 +307,7 @@ export const buildCatalogedSchema = (service: string, serviceType: ServiceType, 
 
             const args = {};
             for (const a of f.arguments) {
-              const { field, dataType } = findDataType(a);
+              const { field, dataType } = findDataType(a, sideEffect);
               if (dataType) Object.assign(args, field);
             }
             if (Object.keys(args).length > 0) {
