+++ packages/gateway-lib/src/utils/index.ts
@@ -10,3 +10,4 @@ export * from './catchResolverErrors';
 export * from './createRemoteData';
 export * from './buildRedisOptions';
 export * from './createGatewayWithAuth0';
+export * from './aclService';
